+++ lib/src/execution_context.dart
@@ -17,8 +17,8 @@ class ExecutionContext {
   /// The response object used to send data back from the function.
   ExecutionResponse get res => ExecutionResponse._(_context.res);
 
-  /// The headers associated with the function execution.
-  ExecutionHeaders get headers => ExecutionHeaders._(_context.req.headers);
+  /// The request headers associated with the function execution.
+  RequestHeaders get headers => RequestHeaders._(_context.req.headers);
 
   /// Logs a message to the execution log.
   void log(String message) => _context.log(message);
@@ -47,10 +47,11 @@ class ExecutionRequest {
   /// Returns the raw binary body of the request.
   List<int> get bodyBinary => _req.bodyBinary;
 
-  /// Key-value pairs of all request headers.
-  ///
-  /// The keys for the headers are in lowercase.
+  /// Raw request headers map (lowercased keys).
   Map<String, dynamic> get headers => _req.headers;
+
+  /// The request headers as a [RequestHeaders] object.
+  RequestHeaders get requestHeaders => RequestHeaders._(_req.headers);
 
   /// The scheme of the request, such as 'http' or 'https'.
   ///
@@ -95,81 +96,108 @@ class ExecutionResponse {
   final dynamic _res;
   const ExecutionResponse._(this._res);
 
+  ///Sends a response with a code 204 No Content status.
   dynamic empty() {
     return _res.empty();
   }
 
-  dynamic json(final Map<String, dynamic> json) {
-    return _res.json(json);
+  /// Converts the data into a JSON string and sets the content-type header to [application/json] with [statusCode].
+  dynamic json(final Map<String, dynamic> json, [int statusCode = 200]) {
+    return _res.json(json, statusCode);
   }
 
-  dynamic binary(Uint8List content) {
-    return _res.binary(content);
+  ///Packages binary bytes, the status code, and the headers into an object with [statusCode].
+  dynamic binary(Uint8List content, [int statusCode = 200]) {
+    return _res.binary(content, statusCode);
   }
 
-  dynamic redirect(String url) {
-    return _res.redirect(url, 301);
+  ///Redirects the client to the specified URL link with [statusCode].
+  dynamic redirect(String url, [int statusCode = 301]) {
+    return _res.redirect(url, statusCode);
   }
 
-  dynamic html(String html) {
-    return _res.text(html, 200, {'content-type': 'text/html'});
+  /// Sends an HTML response with the content-type header set to [text/html] with [statusCode].
+  dynamic html(String html, [int statusCode = 200]) {
+    return _res.text(html, statusCode, {'content-type': 'text/html'});
   }
 
-  dynamic text(String text) {
-    return _res.text(text);
+  ///Converts the body using UTF-8 encoding into a binary Buffer and sends it with [statusCode].
+  dynamic text(String text, [int statusCode = 200]) {
+    return _res.text(text, statusCode);
   }
 
-  dynamic success([String message = '']) {
-    return _res.text(message, 200);
+  /// Sends a success response with an optional message and [statusCode].
+  dynamic success({String message = '', int statusCode = 200}) {
+    return _res.text(message, statusCode);
   }
 
-  dynamic error([String message = '']) {
-    return _res.text(message, 500);
+  /// Sends an error response with an optional message and [statusCode].
+  dynamic error({String message = '', int statusCode = 500}) {
+    return _res.text(message, statusCode);
   }
 }
 
-class ExecutionHeaders {
-  final Map<String, dynamic> headers;
-  const ExecutionHeaders._(this.headers);
-
-  operator [](String key) => headers[key];
+/// Execution Request headers
+///
+/// All keys are lowercase.
+class RequestHeaders {
+  final Map<String, dynamic> _headers;
+  const RequestHeaders._(this._headers);
 
   /// Describes how the function execution was invoked.
   ///
   /// Possible values are `http`, `schedule`, or `event`.
-  String get trigger => headers['x-appwrite-trigger'];
+  String get trigger => _headers['x-appwrite-trigger'];
 
   /// If the function execution was triggered by an event, this describes the triggering event.
-  dynamic get event => headers['x-appwrite-event'];
+  dynamic get event => _headers['x-appwrite-event'];
 
   /// The dynamic API key used for server authentication.
   /// https://appwrite.io/docs/products/functions/develop#dynamic-api-key
-  String? get key => headers['x-appwrite-key'];
+  String? get key => _headers['x-appwrite-key'];
 
   /// If the function execution was invoked by an authenticated user, this is the user's ID.
   ///
   /// This will be `null` for executions triggered by the Appwrite Console or API keys.
-  String? get userId => headers['x-appwrite-user-id'];
+  String? get userId => _headers['x-appwrite-user-id'];
 
   /// The JWT token generated from the invoking user's session.
   ///
   /// This is used to authenticate Server SDKs to respect user access permissions.
-  String? get userJwt => headers['x-appwrite-user-jwt'];
+  String? get userJwt => _headers['x-appwrite-user-jwt'];
 
   /// The country code of the configured locale.
-  String? get countryCode => headers['x-appwrite-country-code'];
+  String? get countryCode => _headers['x-appwrite-country-code'];
 
   /// The continent code of the configured locale.
-  String? get continentCode => headers['x-appwrite-continent-code'];
+  String? get continentCode => _headers['x-appwrite-continent-code'];
 
   /// Describes if the configured locale is within the EU.
   ///
   /// The value will be a string, such as 'true' or 'false'.
-  String? get continentEu => headers['x-appwrite-continent-eu'];
+  String? get continentEu => _headers['x-appwrite-continent-eu'];
 
   /// The IP address of the client that triggered the execution.
-  String? get clientIp => headers['x-appwrite-client-ip'];
+  String? get clientIp => _headers['x-appwrite-client-ip'];
 
   /// The unique ID of the current function execution.
-  String get executionId => headers['x-appwrite-execution-id'];
+  String get executionId => _headers['x-appwrite-execution-id'];
+
+  T requireValue<T>(String key) {
+    final value = _headers[key];
+    if (value is! T) {
+      throw Exception(
+          'Expected header "$key" to be of type $T but got ${value.runtimeType}');
+    }
+    return value;
+  }
+
+  operator [](String key) => _headers[key];
+  Iterable<String> get keys => _headers.keys;
+  Iterable<dynamic> get values => _headers.values;
+  Iterable<MapEntry<String, dynamic>> get entries => _headers.entries;
+  bool containsKey(String key) => _headers.containsKey(key);
+  int get length => _headers.length;
+  forEach(void Function(String key, dynamic value) action) =>
+      _headers.forEach(action);
 }