+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## 1.0.0
+# 1.0.1
+
+- Add optional HTTP status code parameters to ExecutionResponse helpers
+- Add RequestHeaders.requireValue<T>(key) helper to validate headers and enforce types.
+- Lower minimum Dart SDK constraint to Dart 3.0.
+
+# 1.0.0
 
 - Initial release